Feature/pym0002   data leakage fixing (#5)
* PYM0002 - introducing TransactionDto
* PYM0002 - type correction for balance, amount, application.yml fix
* PYM0002 - first validation added
* PYM0002 - usage of NotNull, NotBlank, Size, PositiveOrZero

+++ src/main/java/eu/brevissimus/payment/model/entity/Account.java
@@ -2,7 +2,7 @@ package eu.brevissimus.payment.model.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.util.List;
 
 import jakarta.persistence.CascadeType;
@@ -45,7 +45,7 @@ public class Account extends Auditable {
   private String currency;
 
   @Column(name = "balance")
-  private BigInteger balance;
+  private BigDecimal balance;
 
   @Column(name = "active")
   private boolean active;
